first part - reset resource

+++ src/main/java/org/example/Turn.java
@@ -3,5 +3,5 @@ package org.example;
 public class Turn {
     public int TMt;
     public int TXt;
-    public float TRt;
+    public int TRt;
 }
